chore: change directory structure - lib/prisma/src for prisma module & service + prisma/* for schema + move prisma repository to src/posts
