+++ lib/prisma/src/prisma.module.ts
@@ -1,9 +1,8 @@
 import { Module } from '@nestjs/common';
 import { PrismaService } from './prisma.service';
-import { PostsRepository } from './prisma.repository';
 
 @Module({
-  providers: [PrismaService, PostsRepository],
-  exports: [PostsRepository],
+  providers: [PrismaService],
+  exports: [PrismaService],
 })
 export class PrismaModule {}
