+++ src/posts/posts.module.ts
@@ -1,11 +1,12 @@
 import { Module } from '@nestjs/common';
 import { PostsService } from './posts.service';
 import { PostsController } from './posts.controller';
-import { PrismaModule } from 'prisma/prisma.module';
+import { PrismaModule } from 'lib/prisma/src/prisma.module';
+import { PostsRepository } from './posts.repository';
 
 @Module({
   imports: [PrismaModule],
   controllers: [PostsController],
-  providers: [PostsService],
+  providers: [PostsService, PostsRepository],
 })
 export class PostsModule {}
