+++ src/posts/posts.service.ts
@@ -6,7 +6,7 @@ import {
 import { CreatePostDto } from './dto/create-post.dto';
 import { UpdatePostDto } from './dto/update-post.dto';
 import { Post } from './entities/post.entity';
-import { PostsRepository } from 'prisma/prisma.repository';
+import { PostsRepository } from './posts.repository';
 
 @Injectable()
 export class PostsService {
